Initial Commit
This project was from another repo called "\Unreal-Engine-Battery-Collector\" Turns out the mistake was that you have to start a CPP project. The project in the aforementioned repo was started in blueprint by mistake, and cannot proceed beyone tutorial no. 9 "C++ Battery Collector: Extending the Character Class"

This repo is an attempt to migrate the project to a CPP version by creating a fresh project and migrating everything over. The CPP codes was migrated by creating new corresponding classes in the new project and manually copying the code over file by file. It has not been checked for compile at this commit.